+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 > Read detailed change logs in https://github.com/sakurawald/fuji/commits/dev/
 
+- [core] feature: now will `validate` the `language files` when loading them.
+  - feature: add `argument arity validator`, to check if the `argument number` for each `language value` is matched, and try to fix it if mismatched.
 - [core] feature: enhance the `data migrator`, to provide better cross-version data schema migration.
 - [core] feature: auto-save feature for data files has been updated to save `every 10 seconds` instead of `every minute`.